requirements.txt
# Core NLP (optional — falls back to stdlib if absent)
nltk>=3.8

# Web framework
flask>=3.0

# Config
PyYAML>=6.0

# ML stack (required for v2 ML scorer)
scikit-learn>=1.3
numpy>=1.24
scipy>=1.10
joblib>=1.3

# Testing
pytest>=8.0
